Update app version and refactor MainViewModel classes
Updated the application version in `Package.appxmanifest` from `1.0.0.0` to `1.0.1.0`. Refactored the `Name` and `Address` classes in `MainViewModel.cs`. Removed properties `maiden`, `suffix`, `givenName`, `middleName`, `surname`, `latitude`, and `longitude`. Added new properties `FirstName`, `MiddleName`, `Surname`, `BirthDate`, `Id`, and `IsMarried`. Also, updated the condition for setting the `IsPreviewAvailable` property in the `GenerateXamlForClass` method to check if the `ForMAUI` property is not true.